requirements.txt
black==23.9.1
pytest==7.4.2
pylint==3.0.1
isort==5.12.0
hydra-core==1.3.2